+++ src/components/SearchItems/stylesSearchCharacter.js
