+++ src/components/SearchItems/stylesSearchComics.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import { SearchWriper } from "./stylesSearchCharacter";
+
+const SearchComicsWriper = styled(SearchWriper)`
+    margin: 50px 0 0 0;
+    width: 100%;
+`;
+
+export { SearchComicsWriper }
