+++ src/services/ApiService/ApiService.js
@@ -60,7 +60,7 @@ export function useMarvelService (startLoading = true) {
         const result = await getResource(`${_baseHttp}/comics?apikey=${_apiKey}&titleStartsWith=${title}&limit=${count}&offset=${offset}`);
         return {
             count: result.data.total,
-            data: result.data.results.map(_transformCharacter)
+            data: result.data.results.map(_transformComics)
         } 
     }
